refactor(#15): BookInfoBox 컴포넌트 코드 위치 변경
