+++ src/components/molecules/BookInfoBox/style.ts
